paykit: expose resetPayment function
